页面修改

+++ HuoBanMall/Classes/DataStatistics(数据统计)/controller/HTDataStatisViewController.h
@@ -21,6 +21,6 @@
 /**
  *  标记
  */
-@property (nonatomic, assign) int selectIndex;
+@property (nonatomic, assign) NSInteger selectIndex;
 
 @end
